+++ config-overrides.js
@@ -5,8 +5,8 @@
  * @see https://alchemy.com/blog/how-to-polyfill-node-core-modules-in-webpack-5
  */
 import webpack from 'webpack'
-import { createRequire } from 'module';
-const require = createRequire(import.meta.url);
+import { createRequire } from 'module'
+const require = createRequire(import.meta.url)
 
 const PURE_ESM_MODULES = [
   'ipfs-geoip'
@@ -100,52 +100,56 @@ function webpackOverride (config) {
   return config
 }
 
-export default {
+const configOverride = {
   webpack: webpackOverride,
   jest: (config) => {
-    console.log('config.globals: ', config.globals)
-    console.log('config.roots: ', config.roots)
-    console.log('config.modulePaths: ', config.modulePaths)
-    console.log('config.moduleDirectories: ', config.moduleDirectories)
-    console.log('config.transform: ', config.transform)
-    console.log('config.transformIgnorePatterns: ', config.transformIgnorePatterns)
+    /**
+     * @type {import('jest').Config}
+     */
     return ({
       ...config,
-      globals: {
-        ...config.globals,
-        crypto: require('crypto')
-      },
-      modulePaths: [
-        ...config.modulePaths,
-        'node_modules',
-        '<rootDir>',
-        '<rootDir>/node_modules'
-      ],
-      transform: {
-        // '^.*(?:kubo-rpc-client|ipfsd-ctl).*$': '<rootDir>/jest-esm-transformer.js',
-        '^.+kubo-rpc-client.+$': '<rootDir>/jest-esm-transformer.js',
-        '@multiformats/multiaddr': '<rootDir>/jest-esm-transformer.js',
-        // '^.+ipfsd-ctl/node_modules.+$': '<rootDir>/jest-esm-transformer.js',
-        // '^.+execa/node_modules.+$': '<rootDir>/jest-esm-transformer.js',
-        // 'node_modules/nanoid': '<rootDir>/jest-esm-transformer.js',
-        // '^.+/node_modules/kubo-rpc-client.+$': 'jest-esm-transformer',
-        // '^.+/node_modules/ipfsd-ctl.+$': 'jest-esm-transformer',
-        // '^.+/node_modules/ipfs-core-utils/src/multibases.js$': '<rootDir>/jest-esm-transformer.js',
-        ...config.transform
-        // 'node_modules/kubo-rpc-client/.+\\.(js|jsx|mjs|cjs|ts|tsx)$': 'babel-jest'
-        // 'kubo-rpc-client': 'jest-esm-transformer'
-      },
-      transformIgnorePatterns: [
-        '<rootDir>[/\\\\]node_modules[/\\\\](?!kubo-rpc-client|ipfsd-ctl|ipfsd-ctl\\/node_modules\\/|@libp2p\\/logger|nanoid|temp-write|@multiformats/multiaddr|is-ip|ip-regex|execa|strip-final-newline|npm-run-path|path-key|onetime|mimic-fn|human-signals|is-stream|p-wait-for).+\\.(js|jsx|mjs|cjs|ts|tsx)$',
-        '^.+\\.module\\.(css|sass|scss)$'
-      ],
-      moduleNameMapper: {
-        'ipfsd-ctl': '<rootDir>/node_modules/ipfsd-ctl/dist/src/index.js',
-        '@libp2p/logger': '<rootDir>/node_modules/@libp2p/logger/dist/src/index.js',
-        nanoid: '<rootDir>/node_modules/nanoid/index.js',
-        // '@multiformats/multiaddr': '<rootDir>/node_modules/@multiformats/multiaddr/dist/src/index.js',
-        'ipfs-core-utils/multibases': '<rootDir>/node_modules/ipfs-core-utils/src/multibases.js'
-      }
+      transform: {},
+      setupFilesAfterEnv: ['./src/setupJestTests.js'],
+      setupFiles: ['fake-indexeddb/auto']
+    //   globals: {
+    //     ...config.globals,
+    //     crypto: require('crypto')
+    //   },
+    //   modulePaths: [
+    //     ...config.modulePaths,
+    //     'node_modules',
+    //     '<rootDir>',
+    //     '<rootDir>/node_modules'
+    //   ],
+    //   transform: {
+    //     // '.js': 'jest-esm-transformer',
+    //     //   // '^.*(?:kubo-rpc-client|ipfsd-ctl).*$': '<rootDir>/jest-esm-transformer.js',
+    //     // '^.+kubo-rpc-client.+$': '<rootDir>/jest-esm-transformer.js',
+    //     '^.+kubo-rpc-client.+$': 'jest-esm-transformer',
+    //     //   // '@multiformats/multiaddr': '<rootDir>/jest-esm-transformer.js',
+    //     //   // '^.+ipfsd-ctl/node_modules.+$': '<rootDir>/jest-esm-transformer.js',
+    //     //   // '^.+execa/node_modules.+$': '<rootDir>/jest-esm-transformer.js',
+    //     //   // 'node_modules/nanoid': '<rootDir>/jest-esm-transformer.js',
+    //     //   // '^.+/node_modules/kubo-rpc-client.+$': 'jest-esm-transformer',
+    //     //   // '^.+/node_modules/ipfsd-ctl.+$': 'jest-esm-transformer',
+    //     //   // '^.+/node_modules/ipfs-core-utils/src/multibases.js$': '<rootDir>/jest-esm-transformer.js',
+    //     ...config.transform
+    //   //   // 'node_modules/kubo-rpc-client/.+\\.(js|jsx|mjs|cjs|ts|tsx)$': 'babel-jest'
+    //   //   // 'kubo-rpc-client': 'jest-esm-transformer'
+    //   }
+    //   // transformIgnorePatterns: [
+    //   //   '<rootDir>[/\\\\]node_modules[/\\\\](?!kubo-rpc-client|ipfsd-ctl|ipfsd-ctl\\/node_modules\\/|@libp2p\\/logger|nanoid|temp-write|@multiformats/multiaddr|is-ip|ip-regex|execa|strip-final-newline|npm-run-path|path-key|onetime|mimic-fn|human-signals|is-stream|p-wait-for).+\\.(js|jsx|mjs|cjs|ts|tsx)$',
+    //   //   '^.+\\.module\\.(css|sass|scss)$'
+    //   // ],
+    //   // moduleNameMapper: {
+    //   //   'ipfsd-ctl': '<rootDir>/node_modules/ipfsd-ctl/dist/src/index.js',
+    //   //   '@libp2p/logger': '<rootDir>/node_modules/@libp2p/logger/dist/src/index.js',
+    //   //   nanoid: '<rootDir>/node_modules/nanoid/index.js',
+    //   //   // '@multiformats/multiaddr': '<rootDir>/node_modules/@multiformats/multiaddr/dist/src/index.js',
+    //   //   'ipfs-core-utils/multibases': '<rootDir>/node_modules/ipfs-core-utils/src/multibases.js'
+    //   // }
     })
   }
 }
+
+export default configOverride
