+++ package.json
@@ -23,7 +23,7 @@
     "build:darwin:linux": "cross-env REACT_APP_GIT_REV=`git rev-parse --short HEAD` react-app-rewired-esm build",
     "build:icons": "svgr -d src/icons node_modules/ipfs-css/icons && standard --fix src/icons/*",
     "test": "run-s -cl test:unit test:build test:e2e",
-    "test:unit": "react-app-rewired-esm test --env=jsdom --runInBand --watchAll=false",
+    "test:unit": "cross-env NODE_OPTIONS=--experimental-vm-modules react-app-rewired-esm test --env=jsdom --runInBand --watchAll=false",
     "test:unit:coverage": "react-app-rewired-esm test  --env=jsdom --runInBand --watchAll=false --coverageDirectory='./.nyc_output' --coverage",
     "test:unit:watch": "react-app-rewired-esm test --env=jsdom",
     "test:e2e": "cross-env REACT_APP_ENV=test npx playwright test -c ./test/e2e",
@@ -147,6 +147,7 @@
     "babel-eslint": "^10.1.0",
     "babel-jest": "^29.3.1",
     "babel-plugin-istanbul": "^6.1.1",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
     "basic-auth": "^2.0.1",
     "big.js": "^5.2.2",
     "bundlesize": "0.18.1",
@@ -166,7 +167,7 @@
     "eslint-plugin-standard": "^4.1.0",
     "eslint-plugin-storybook": "^0.6.1",
     "esm": "^3.2.25",
-    "fake-indexeddb": "^3.1.2",
+    "fake-indexeddb": "^3.1.8",
     "get-port": "^5.1.1",
     "go-ipfs": "0.16.0",
     "http-proxy": "^1.18.1",
