+++ src/bundles/analytics.test.js
@@ -1,4 +1,5 @@
-/* global describe, it, expect, beforeEach, afterEach, jest */
+/* global describe, it, expect, beforeEach, afterEach */
+import { jest } from '@jest/globals'
 import { composeBundlesRaw } from 'redux-bundler'
 import createAnalyticsBundle from './analytics.js'
