+++ src/bundles/files.test.js
@@ -1,4 +1,5 @@
-/* global it, expect, jest */
+/* global it, expect */
+import { jest } from '@jest/globals'
 import { composeBundlesRaw } from 'redux-bundler'
 import createFilesBundle from './files/index.js'
 import CID from 'cids'
