+++ src/setupTests.js
@@ -1,4 +1,5 @@
-import { configure } from 'enzyme'
+import * as enzyme from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
 
+const { configure } = enzyme.default
 configure({ adapter: new Adapter() })
